+++ pacman_ui_fx/src/main/java/de/amr/games/pacman/ui/fx/_3d/entity/World3D.java
@@ -31,7 +31,7 @@ import static de.amr.games.pacman.model.common.world.World.TS;
 
 import java.util.stream.Stream;
 
-import de.amr.games.pacman.lib.math.Vector2d;
+import de.amr.games.pacman.lib.math.Vector2f;
 import de.amr.games.pacman.model.common.GameModel;
 import de.amr.games.pacman.model.common.actors.Ghost;
 import de.amr.games.pacman.model.common.actors.GhostState;
@@ -143,16 +143,16 @@ public class World3D extends Group {
 		}
 	}
 
-	private boolean isAccessGranted(Stream<Ghost> ghosts, Vector2d doorPosition) {
+	private boolean isAccessGranted(Stream<Ghost> ghosts, Vector2f doorPosition) {
 		return ghosts.anyMatch(ghost -> isAccessGranted(ghost, doorPosition));
 	}
 
-	private boolean isAccessGranted(Ghost ghost, Vector2d doorPosition) {
+	private boolean isAccessGranted(Ghost ghost, Vector2f doorPosition) {
 		return ghost.isVisible() && ghost.is(RETURNING_TO_HOUSE, ENTERING_HOUSE, LEAVING_HOUSE)
 				&& inDoorDistance(ghost, doorPosition);
 	}
 
-	private boolean inDoorDistance(Ghost ghost, Vector2d doorPosition) {
+	private boolean inDoorDistance(Ghost ghost, Vector2f doorPosition) {
 		return ghost.position().euclideanDistance(doorPosition) <= (ghost.is(LEAVING_HOUSE) ? TS : 3 * TS);
 	}
 }
